исправлено обработки ошибок и валидации в ItemServiceImpl и BookingServiceImp

+++ src/main/java/ru/practicum/shareit/user/UserServiceImpl.java
@@ -80,6 +80,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void deleteUser(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new NotFoundException("Пользователь с ID " + userId + " не найден");
+        }
         userRepository.deleteById(userId);
     }
 }
